move slides to pages
