+++ pages/asset-manager/slides/howIncentivised.js
