+++ pages/asset-manager/slides/index.js
