+++ pages/list-asset/slides/assetLocation.js
