+++ pages/list-asset/slides/availableAssets.js
