+++ pages/list-asset/slides/collateral.js
