+++ pages/list-asset/slides/confirm.js
