+++ pages/list-asset/slides/confirmDesktop.js
