+++ pages/onboarding/slides/benefits.js
