+++ pages/onboarding/slides/index.js
