+++ pages/onboarding/slides/invest.js
