+++ pages/onboarding/slides/investmentFund.js
