+++ pages/onboarding/slides/nextGeneration.js
