+++ pages/onboarding/slides/requiredSetup.js
