+++ pages/onboarding/slides/smartContracts.js
