+++ pages/onboarding/slides/welcome.js
